ME46055_2021_36.mlx
Contains a Matlab live script of the model. Symbolic variables have been used to derive the mass stiffness and damping matrices of the system. The equilibrium position of all displacements is 0 and the arm is considered to have a 90-degree orientation at rest position at the elbow. Eigenmodes and frequencies have been obtained from the resulting matrices and have been used to calculate the damping ratios.